requirements.txt
agno
openai
anthropic
sqlalchemy
httpx
python-dotenv
ddgs
fastapi
jupyter
